+++ tests/testthat/test-TotalCopheneticIndex.R
@@ -22,7 +22,7 @@ test_that("Trees from Mir et al. 2013 are scored correctly", {
   nasty <- structure(list(edge = structure(
     c(9, 12, 10, 13, 11, 10, 11, 13, 10, 13, 12, 9,
       5, 10,  1,  2,  3, 13,  9,  4, 11,  7,  8, 6),
-    .Dim = c(12, 2)),
+    dim = c(12, 2)),
     Nnode = 5L,
     tip.label = letters[1:8]),
     class = "phylo")