+++ tests/testthat/test-tree_write.R
@@ -11,7 +11,7 @@ test_that("Write is successful", {
   nasty <- structure(list(edge = structure(
     c(9, 12, 10, 13, 11, 10, 11, 13, 10, 13, 12, 9,
       5, 10,  1,  2,  3, 13,  9,  4, 11,  7,  8, 6),
-    .Dim = c(12, 2)),
+    dim = c(12, 2)),
     Nnode = 5L,
     tip.label = 0:7),
     class = "phylo") # Danger: Do not plot!